RadicalStroke parse CJKRadicals.txt directly, not via UnicodeMap
